requirements.txt
Django==2.2.3
django-environ==0.4.5
djangorestframework==3.9.4
django-cors-headers==3.1.0
markdown==3.1.1
django-filter==2.1.0
pytz==2019.1
setuptools==41.0.1
sqlparse==0.3.0
mysqlclient==1.4.2
pyjwt==1.7.1